Fixed a lot of bugs. Refactored *_Inotify_Event
Removed the dependency between Event and Watch in Event::fullpath()
method. From now on Event is just a stupid data container - as planned
before. +Many other minor refactorings and bug fixes.

Currently the test suite shows no errors but there are still design
problems: Creating events and assigning the fullpath to the constructor
requires the parent watch to exist. I'm not sure if this is guranteed.
Will have to investigate and maybe rewrite some code.