Removed the IN_*_FOLLOW option. Refactored code
The IN_X_RECURSIVE_FOLLOW options was meant to ... that in recursive
mode new directories should being watched automatically. I think this
behaviour is desired by default so I removed that needless option to
make things not to complicated. Maybe it will be added again once? ;)

Removed the method Jm_Os_Inotify::watchRecursive().
Have generalized Jm_Os_Inotify::watch() so that it can do the work with
almost the same effort while not looking ugly because of this.